Some progress, need to rethink if I want positions to be x/y, or a delta from the start of the buffer; Move the UITextInput support to a different file
