+++ Sources/SwiftTerm/iOS/iOSTerminalView.swift
@@ -32,7 +32,7 @@ import CoreGraphics
  * Use the `configureNativeColors()` to set the defaults colors for the view to match the OS
  * defaults, otherwise, this uses its own set of defaults colors.
  */
-open class TerminalView: UIScrollView, UITextInputTraits, UIKeyInput, UIScrollViewDelegate, UITextInput {
+open class TerminalView: UIScrollView, UITextInputTraits, UIKeyInput, UIScrollViewDelegate {
     
     struct FontSet {
         public let normal: UIFont
@@ -92,6 +92,8 @@ open class TerminalView: UIScrollView, UITextInputTraits, UIKeyInput, UIScrollVi
     var colors: [UIColor?] = Array(repeating: nil, count: 256)
     var trueColors: [Attribute.Color:UIColor] = [:]
     var transparent = TTColor.transparent ()
+    var _markedTextRange: TerminalTextRange?
+    public var tokenizer: UITextInputTokenizer = UITextInputStringTokenizer () // TerminalInputTokenizer()
 
     var fontSet: FontSet
     /// The font to use to render the terminal
@@ -137,7 +139,7 @@ open class TerminalView: UIScrollView, UITextInputTraits, UIKeyInput, UIScrollVi
     func setupUIKitSelection ()
     {
         let initialPosition = TerminalTextPosition (Position (col: 0, row: 0))
-        _selectionTextRange = TerminalTextRange(start: initialPosition, end: initialPosition)
+        _selectionTextRange = nil
     }
     
     func setupDisplayUpdates ()
@@ -741,323 +743,6 @@ open class TerminalView: UIScrollView, UITextInputTraits, UIKeyInput, UIScrollVi
     public override func pressesEnded(_ presses: Set<UIPress>, with event: UIPressesEvent?) {
         // guard let key = presses.first?.key else { return }
     }
-
-    // Classes to support UITextInput and other text input protocols
-    
-    // The text position is relative to the start of the buffer (buffer.yBase)
-    class TerminalTextPosition: UITextPosition {
-        var pos: Position
-        init (_ pos: Position)
-        {
-            self.pos = pos
-        }
-        public override var debugDescription: String {
-            get {
-                return "(col=\(pos.col),row=\(pos.row)"
-            }
-        }
-    }
-    
-    class TerminalTextRange: UITextRange {
-        var _start: TerminalTextPosition
-        var _end: TerminalTextPosition
-        
-        init(start: TerminalTextPosition, end: TerminalTextPosition) {
-            _start = start
-            _end = end
-        }
-        
-        override var start: UITextPosition { _start }
-        override var end: UITextPosition { _end }
-        override var isEmpty: Bool { _start == _end }
-        
-        public override var debugDescription: String {
-            get {
-                return "\(start)-\(end)"
-            }
-        }
-    }
-    
-    class TerminalSelectionRect: UITextSelectionRect {
-        var _rect: CGRect
-        var _containsStart: Bool
-        var _containsEnd: Bool
-        
-        override var writingDirection: NSWritingDirection {
-          return .leftToRight
-        }
-        
-        override var isVertical: Bool {
-          return false
-        }
-        
-        override var rect: CGRect {
-          return _rect
-        }
-        
-        override var containsStart: Bool {
-          return _containsStart
-        }
-        
-        override var containsEnd: Bool {
-          return _containsEnd
-        }
-
-        init(rect: CGRect, range: TerminalTextRange, string: String) {
-            _rect = rect
-            
-            print ("THIS IS WRONG:")
-            _containsStart = true
-            _containsEnd = true
-        }
-    }
-    
-    // This code is currently not enabled, we are using a standard string tokenizer, see UITextInputStringTokenizer below
-    class TerminalInputTokenizer: NSObject, UITextInputTokenizer {
-        func pabort (_ msg: String)
-        {
-            print (msg)
-            abort()
-        }
-        func rangeEnclosingPosition(_ position: UITextPosition, with granularity: UITextGranularity, inDirection direction: UITextDirection) -> UITextRange? {
-            pabort ("PROTO: MIT/Range")
-
-            return nil
-        }
-
-        func isPosition(_ position: UITextPosition, atBoundary: UITextGranularity, inDirection: UITextDirection) -> Bool {
-            pabort ("PROTO: MIT/offset")
-            return false
-        }
-
-        func position(from position: UITextPosition, toBoundary granularity: UITextGranularity, inDirection direction: UITextDirection) -> UITextPosition? {
-            pabort ("PROTO: MIT/position1")
-            return nil
-        }
-
-        func isPosition(_ position: UITextPosition, withinTextUnit granularity: UITextGranularity, inDirection direction: UITextDirection) -> Bool {
-            pabort ("PROTO: MIT/position")
-            return false
-        }
-    }
-    
-    /// Confromance to UITextInput
-    func pabort (_ msg: String)
-    {
-        print (msg)
-        abort ()
-    }
-
-    public func text(in range: UITextRange) -> String? {
-        let termRange = range as! TerminalTextRange
-        let start = (termRange.start as! TerminalTextPosition).pos
-        let end = (termRange.end as! TerminalTextPosition).pos
-        
-        let text = terminal.getText(start: start, end: end)
-        
-        print ("text(in range: \(start)-\(end)) -> \"\(text)\"")
-        return text
-    }
-
-    public func replace(_ range: UITextRange, withText text: String) {
-        guard let r = range as? TerminalTextRange else {
-            pabort ("replace: Called with a non TerminalTextRange")
-            return
-        }
-        if text == "" {
-            return
-        }
-        print ("REPLACE: found a replacement with body: \(text)")
-    }
-
-    public var selectedTextRange: UITextRange? {
-        get {
-            
-            // TODO: rather than creating this every time, create it when the selection
-            // is changed, and always update _selectionTextRange
-            // This is temporary while I get the other methods working
-            if selection.active && selection.hasSelectionRange {
-                return TerminalTextRange (start: TerminalTextPosition(selection.start), end: TerminalTextPosition(selection.end))
-            }
-            return _selectionTextRange
-        }
-        set {
-            if let newRange = newValue as? TerminalTextRange {
-                if let start = (newRange.start as? TerminalTextPosition),
-                    let end = (newRange.end as? TerminalTextPosition) {
-
-                    selection.setSelection(start: start.pos, end: end.pos)
-                }
-            } else {
-                
-                selection.active = false
-            }
-        }
-    }
-
-    var _markedTextRange: TerminalTextRange = TerminalTextRange (start: TerminalTextPosition (Position (col: 0, row: 0)), end: TerminalTextPosition (Position (col: 0, row: 0)))
-    
-    public var markedTextRange: UITextRange? {
-        get {
-            //pabort ("Request for marked-text-range")
-            print ("Request for marked-text-range")
-            return _markedTextRange
-        }
-    }
-
-    public var markedTextStyle: [NSAttributedString.Key : Any]? {
-        get {
-            pabort ("PROTO: markedTextStyle")
-            return nil
-        }
-        set {
-            pabort ("PROTO: set markedTextStyle")
-        }
-    }
-
-    public func setMarkedText(_ markedText: String?, selectedRange: NSRange) {
-        pabort ("PROTO: etMarkedText")
-    }
-
-    public func unmarkText() {
-        pabort ("PROTO: unmarktext")
-    }
-
-    public var beginningOfDocument: UITextPosition {
-        get {
-            return TerminalTextPosition(Position (col: 0, row: 0))
-        }
-    }
-
-    public var endOfDocument: UITextPosition {
-        get {
-            return TerminalTextPosition(Position (col: terminal.buffer.cols, row: terminal.buffer.lines.count))
-        }
-    }
-
-    public func beginFloatingCursor(at: CGPoint) {
-        print ("oo")
-    }
-    public func textRange(from fromPosition: UITextPosition, to toPosition: UITextPosition) -> UITextRange? {
-        guard let from = fromPosition as? TerminalTextPosition, let to = toPosition as? TerminalTextPosition else {
-            fatalError()
-        }
-        print("[Geometry] form range [\(from.pos) ..< \(to.pos)]")
-        
-        return TerminalTextRange(start: from, end: to)
-    }
-
-    public func position(from position: UITextPosition, offset: Int) -> UITextPosition? {
-        guard let pos = position as? TerminalTextPosition else {
-            abort ()
-        }
-        print ("POSITION-offset: \(pos.pos) \(offset)")
-        let p = (position as! TerminalTextPosition).pos
-        var col = p.col + offset
-        col = min (max (col, 0), terminal.cols-1)
-        print ("POSITION-offset: \(pos.pos) \(offset) going to-> \(col)")
-        return TerminalTextPosition (Position (col: col, row: p.row))
-        return nil
-    }
-
-    public func position(from position: UITextPosition, in direction: UITextLayoutDirection, offset: Int) -> UITextPosition? {
-        pabort ("PROTO: position2")
-        return nil
-    }
-
-    public func compare(_ position: UITextPosition, to other: UITextPosition) -> ComparisonResult {
-        if let a = position as? TerminalTextPosition {
-            if let b = other as? TerminalTextPosition {
-                switch Position.compare(a.pos, b.pos){
-                case .before:
-                    return .orderedAscending
-                case .after:
-                    return .orderedDescending
-                case .equal:
-                    return .orderedSame
-                }
-            }
-        }
-        return .orderedSame
-    }
-
-    public func offset(from: UITextPosition, to toPosition: UITextPosition) -> Int {
-        guard let start = from as? TerminalTextPosition, let end = toPosition as? TerminalTextPosition else {
-          fatalError()
-        }
-        let str = terminal.getText (start: start.pos, end: end.pos)
-        return str.utf16.count
-    }
-
-    public var tokenizer: UITextInputTokenizer = UITextInputStringTokenizer () // TerminalInputTokenizer()
-
-    public func position(within range: UITextRange, farthestIn direction: UITextLayoutDirection) -> UITextPosition? {
-        pabort ("PROTO: position3")
-        return nil
-    }
-
-    public func characterRange(byExtending position: UITextPosition, in direction: UITextLayoutDirection) -> UITextRange? {
-        pabort ("PROTO: characterRnage")
-        return nil
-    }
-
-    public func baseWritingDirection(for position: UITextPosition, in direction: UITextStorageDirection) -> NSWritingDirection {
-        return .leftToRight
-    }
-
-    public func setBaseWritingDirection(_ writingDirection: NSWritingDirection, for range: UITextRange) {
-        pabort ("PROTO: setBaseWritingDirection")
-
-    }
-
-    public func firstRect(for range: UITextRange) -> CGRect {
-        guard let r = range as? TerminalTextRange else {
-            pabort ("firstRect (for range: UITextRange) received a non-TerminalTextRange")
-            return CGRect.zero
-        }
-        print ("TODO: firstRect (for Range) needs SCROLLSUPPORT + CORRECTREGION)")
-        return bounds
-    }
-
-    public func caretRect(for position: UITextPosition) -> CGRect {
-        guard let pos = position as? TerminalTextPosition  else {
-            abort ()
-        }
-        
-        print ("PROTO: caretRect for \(pos.pos)")
-        return bounds
-    }
-
-    // Trigger this by hitting the microphone
-    public func selectionRects(for range: UITextRange) -> [UITextSelectionRect] {
-        guard let myRange = range as? TerminalTextRange else {
-            print ("FATAL/PROTO: selectionRects does not get a TerminalTextRange")
-            abort ()
-            return []
-        }
-        print ("TODO: selectionRects (for Range) needs SCROLLSUPPORT + CORRECTREGION)")
-        return [TerminalSelectionRect(rect: bounds, range: myRange, string: text (in: range) ?? "")]
-    }
-
-    // Trigger this by long-pressing the space-bar
-    public func closestPosition(to point: CGPoint) -> UITextPosition? {
-        let col = min (max (0, Int (point.x / cellDimension.width)), terminal.rows)
-        let row = min (max (0, Int (point.y / cellDimension.height)), terminal.cols)
-        
-        // TODO: probably this should return a position offset by the scroll position
-        return TerminalTextPosition (Position(col: col, row: row))
-    }
-
-    public func closestPosition(to point: CGPoint, within range: UITextRange) -> UITextPosition? {
-        pabort ("PROTO: closestPosition")
-        return nil
-    }
-
-    public func characterRange(at point: CGPoint) -> UITextRange? {
-        pabort ("PROTO: characterRange")
-        return nil
-    }
-
 }
 
 extension TerminalView: TerminalDelegate {
